+++ pkg/client/client.go
@@ -152,6 +152,8 @@ func (watcher *Watcher) waitForDeployment(id, appName, version string) error {
 		switch taskInfo.Status {
 		case models.StatusFailedMessage:
 			return fmt.Errorf("The deployment has failed, please check logs.\n%s", taskInfo.StatusReason)
+		case models.StatusCancelledMessage:
+			return fmt.Errorf("The deployment was cancelled because a newer deployment superseded it.\n%s", taskInfo.StatusReason)
 		case models.StatusInProgressMessage:
 			if !isDeploymentOverTime(retryCount, clientConfig.RetryInterval, clientConfig.ExpectedDeploymentTime) {
 				log.Println("Application deployment is in progress...")